+++ .travis.yml
@@ -2,7 +2,7 @@ sudo: false
 matrix:
     fast_finish: true
     include:
-        - python: "3.4.4"
+        - python: "3.4"
           env: TOXENV=py34
         - python: "3.5"
           env: TOXENV=py35
@@ -19,4 +19,5 @@ language: python
 script:
     - python3 test
     - flake8
-after_success: coveralls
+after_success:
+    - coveralls
