+++ Onaeri/lamp.py
