+++ Onaeri/settings/Hallway.py
@@ -1,0 +1,32 @@
+"""
+Basic settings
+"""
+# Time alarm is set in (hour, minute)
+alarmTime = (6, 0)
+# Bedtime in (hour, minute)
+sleepTime = (22, 0)
+
+
+# Allow the lamps to automatically go off in the evening.
+autoPowerOff = False
+# Allow the lamps to automatically go on in the morning.
+autoPowerOn = False
+
+
+# How many minutes before your alarm time should the lamps go on?
+alarmOffset = 0
+# How many minutes before your bedtime should the lamps notify you?
+windDownTime = 30
+
+
+"""
+Advanced settings
+"""
+# Duration of morningslope in minutes
+morningSlopeDuration = 180
+
+# Duration of eveningslope in minutes
+eveningSlopeDuration = 500
+
+# Temporary cycle deviation duration in minutes
+deviationDuration = 0
