+++ Onaeri/settings/Jurrien.py
@@ -1,0 +1,32 @@
+"""
+Basic settings
+"""
+# Time alarm is set in (hour, minute)
+alarmTime = (6, 30)
+# Bedtime in (hour, minute)
+sleepTime = (22, 0)
+
+
+# Allow the lamps to automatically go off in the evening.
+autoPowerOff = True
+# Allow the lamps to automatically go on in the morning.
+autoPowerOn = True
+
+
+# How many minutes before your alarm time should the lamps go on?
+alarmOffset = 30
+# How many minutes before your bedtime should the lamps notify you?
+windDownTime = 30
+
+
+"""
+Advanced settings
+"""
+# Duration of morningslope in minutes
+morningSlopeDuration = 60
+
+# Duration of eveningslope in minutes
+eveningSlopeDuration = 400
+
+# Temporary cycle deviation duration in minutes
+deviationDuration = 45
