+++ Onaeri/settings/dynamic/Hallway.json
@@ -1,0 +1,1 @@
+{"min": {"brightness": 0, "color": 0}, "max": {"brightness": 1000, "color": 1000}, "power": {"off": 2640, "on": 720}, "features": {"color": false, "temp": false, "dim": true}, "meta": {"called": 1516632608.616603, "created": 1515766522.184879}}
