+++ Onaeri/settings/dynamic/Jurrien lamp.json
@@ -1,0 +1,1 @@
+{"min": {"brightness": 0, "color": 0}, "max": {"brightness": 1000, "color": 985}, "power": {"off": 2640, "on": 720}, "features": {"color": false, "temp": true, "dim": true}, "meta": {"called": 1516632608.6198962, "created": 1515766522.184879}}
