+++ Onaeri/settings/dynamic/Sander plafond.json
@@ -1,0 +1,1 @@
+{"min": {"brightness": 0, "color": 0}, "max": {"brightness": 1000, "color": 1000}, "power": {"off": 2820, "on": 1130}, "features": {"color": false, "temp": true, "dim": true}, "meta": {"called": 1516632608.6128309, "created": 1515766522.184879}}
