+++ Onaeri/settings/dynamic/Sander.json
@@ -1,0 +1,1 @@
+{"min": {"brightness": 0, "color": 0}, "max": {"brightness": 1000, "color": 1000}, "power": {"off": 2820, "on": 1130}, "features": {"dim": true, "temp": false, "color": true}, "meta": {"called": 1516632608.615458, "created": 1515776162.758464}}
