+++ Onaeri/settings/dynamic/Template Two.json
@@ -1,0 +1,1 @@
+{"min": {"brightness": 0, "color": 0}, "max": {"brightness": 5, "color": 8}, "power": {"off": 2820, "on": 1020}, "features": {"color": false, "temp": true, "dim": true}, "meta": {"called": 1516632608.7674649, "created": 1516632608.6063669}}
