+++ test/__main__.py
@@ -6,7 +6,8 @@ import shutil
 Create a setting file for the test procedure
 """
 folderPath = os.path.dirname(os.path.abspath(__file__))
-folderPath = os.path.split(folderPath)[0] + "/settings"
+folderPath = os.path.split(folderPath)[0] + "/Onaeri/settings"
+print(folderPath)
 
 srcfile = folderPath + '/Template.py'
 dstfile = folderPath + '/test.py'
