+++ test/test_observer.py
@@ -1,5 +1,5 @@
-from ..observer import Observer
-from ..lamp import Lamp
+from ..Onaeri.observer import Observer
+from ..Onaeri.lamp import Lamp
 
 
 def test_init():
